[CORE][ARM] Supporting transformer model on the ARM CPU

+++ lite/backends/arm/math/topk.cc
@@ -20,13 +20,13 @@ namespace paddle {
 namespace lite {
 namespace arm {
 namespace math {
-bool comp_func(std::pair<float, int> a, std::pair<float, int> b) {
+bool comp_func(std::pair<float, int64_t> a, std::pair<float, int64_t> b) {
   return (a.first > b.first);
 }
 
 void topk(const float* in_data,
           float* out_val,
-          int* out_ind,
+          int64_t* out_ind,
           int m,
           int n,
           int k,
@@ -34,8 +34,8 @@ void topk(const float* in_data,
   for (int i = 0; i < m; i++) {
     const float* in_tmp = in_data + i * n;
     float* out_val_tmp = out_val + i * k;
-    int* out_ind_tmp = out_ind + i * k;
-    std::vector<std::pair<float, int>> vec;
+    int64_t* out_ind_tmp = out_ind + i * k;
+    std::vector<std::pair<float, int64_t>> vec;
     for (int j = 0; j < n; j++) {
       vec.push_back(std::make_pair(in_tmp[j], j));
     }
